refactor(call): move intake and idempotency helpers
