+++ docs/context/architecture/import-boundaries.md
@@ -5,6 +5,7 @@ sources:
   - pyproject.toml
   - .github/workflows/ci.yml
   - src/treg/application/__init__.py
+  - src/treg/application/call/__init__.py
   - src/treg/domain/__init__.py
   - src/treg/domain/governance/__init__.py
   - src/treg/domain/governance/access.py
@@ -54,6 +55,8 @@ import identity but cannot import the API, routers, or application layer. Future
 absent until their packages exist, so no placeholder domain makes a future boundary look active.
 Governance owns shared tool/project ACLs, tag-budget rules, and public-demo rate policy. The package also
 forbids direct FastAPI and Starlette imports; semantic policy errors are translated by each HTTP interface.
+The call application package owns request intake and idempotency state; the remaining call-kernel
+collaborators retain their current owners.
 
 Two direct edges are precise exceptions. `cli.ensure_proxy_dependency` imports `cryptography` only after
 the user invokes the optional proxy feature and offers to install the proxy extra first.
