+++ docs/context/architecture/proxy-model.md
@@ -4,6 +4,8 @@ status: shipped
 sources:
   - src/treg/proxy.py
   - src/treg/api.py
+  - src/treg/application/call/idempotency.py
+  - src/treg/application/call/intake.py
   - src/treg/routers/call.py
 related:
   - architecture/data-model.md
